# auth-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth-service_internal_domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      picture:
        description: Allows NULL
        type: string
      role:
        description: '"user"'
        type: string
      status:
        description: '"unverified", "verified"'
        type: string
      updated_at:
        type: string
    type: object
  auth-service_internal_web_helpers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth-service_internal_web_helpers.OKResponse:
    properties:
      message:
        type: string
    type: object
  internal_auth.EmailLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.GoogleLoginRequest:
    properties:
      id_token:
        type: string
    required:
    - id_token
    type: object
  internal_auth.LoginResult:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth-service_internal_domain.User'
    type: object
  internal_auth.LoginSuccessResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/auth-service_internal_domain.User'
    type: object
  internal_auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.RefreshTokenResponse:
    properties:
      access_token:
        type: string
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_auth.ResendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_auth.SetPasswordRequest:
    properties:
      password:
        minLength: 8
        type: string
    required:
    - password
    type: object
  internal_auth.VerifyRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is the API for the HealthMate application.
  title: HealthMate API
  version: "1.0"
paths:
  /auth/app:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email and password, returns access/refresh
        tokens.
      parameters:
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_auth.EmailLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginSuccessResponse'
      summary: Email/Password Login
      tags:
      - Auth
  /auth/google:
    post:
      consumes:
      - application/json
      description: Authenticates a user using a Google ID token and returns access/refresh
        tokens.
      parameters:
      - description: Google ID Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/internal_auth.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginSuccessResponse'
      summary: Google Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs the user out by invalidating the provided refresh token.
      parameters:
      - description: Refresh Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
      security:
      - BearerAuth: []
      summary: User Logout
      tags:
      - Auth
  /auth/otp/resend:
    post:
      consumes:
      - application/json
      description: Generates and sends a new OTP to the user's email if the account
        is unverified.
      parameters:
      - description: User's Email
        in: body
        name: email_info
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Always returns a success message for security reasons
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "409":
          description: Returned if account is already verified
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      summary: Resend OTP
      tags:
      - Auth
  /auth/otp/verify:
    post:
      consumes:
      - application/json
      description: Verifies OTP and logs the user in by returning tokens.
      parameters:
      - description: Email and OTP
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/internal_auth.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginResult'
        "400":
          description: Invalid OTP or request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: OTP not found or expired
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      summary: Verify Account & Login
      tags:
      - Auth
  /auth/password:
    post:
      consumes:
      - application/json
      description: Sets a password for the currently authenticated user. Requires
        Bearer token.
      parameters:
      - description: New Password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/internal_auth.SetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
      security:
      - BearerAuth: []
      summary: Set User Password
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Provides a new access token if the given refresh token is valid.
      parameters:
      - description: Refresh Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.RefreshTokenResponse'
      security:
      - BearerAuth: []
      summary: Refresh Access Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email, password, and name.
      parameters:
      - description: User registration info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth-service_internal_domain.User'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "409":
          description: email is already registered
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "500":
          description: an unexpected error occurred
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      summary: Register New User
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"